+++ 9-oh-my-rpg/state-meta/src/types.ts
@@ -6,8 +6,8 @@ import { BaseUState } from '@offirmo-private/state'
 
 interface State extends BaseUState {
 
-	// undefined = no created yet
-	// null = opted out
+	// undefined = not created yet
+	// null = opted out DEPRECATED we can always backup, just no social features
 	persistence_id: undefined | string
 
 	is_web_diversity_supporter: boolean