+++ A-apps/the-boring-rpg/state/src/build.ts
@@ -1,5 +1,5 @@
 
 // THIS FILE IS AUTO GENERATED!
-export const VERSION: string = '0.64.3'
-export const NUMERIC_VERSION: number = 0.6403 // for easy comparisons
-export const BUILD_DATE: string = '20200612_00h29'
+export const VERSION: string = '0.64.4'
+export const NUMERIC_VERSION: number = 0.6404 // for easy comparisons
+export const BUILD_DATE: string = '20200618_10h45'