+++ A-apps/the-boring-rpg/client-browser/CHANGELOG.md
@@ -11,6 +11,7 @@
 - 😷 chore: some internal refactors, bump dependencies
 
 ## v0.65.1
+- 🤩 feature: 14 new adventures (now totalling 201!)
 - 😍 feature: 36 new backgrounds (now totalling 134!)
 - 😅 fix: favicon status didn't seem to always work correctly
 - 😷 chore: some internal refactors, bump dependencies